[BUGFIX] Better handling for fields with items configuration
This patch allows to define empty items for fields with items
configuration. This is useful for e.g. selectbox fields as the first
value may be empty to not force any default value.

Furthermore the JavaScript is changed to set the current line number
(key) for items if no value is defined. This is allows to only define
the labels in an items field and use the line numbers as their values.