[TASK] composer.json: remove version field and use branch-alias if needed
The version does not have to be specified in the file (anymore). It's easier to
omit the version field in composer.json. Packagist parses versions from tags
and branch names. New versions of your package are automatically fetched.

Tag/version names should match 'X.Y.Z', or 'vX.Y.Z', with an optional suffix for
RC, beta, alpha or patch versions. So it's fully ok to keep the prefixed "v1.2.3"
syntax.