Fix sorting of fields in an inline type
The drag & drop operaiton in the backend re-orders the form elements,
such that the iteration in the Mask storage update is in the correct
order, but that order is destroyed by sortJson. This fixed sorting is
valuable to version control, so instead this commit introduces an extra
field 'order' in the column of the inline type. This is used in the TCA
generator and before the backend view renders.