Avoid forced sort order by column name in inline elements
The mask.json is sorted to have a consistent file to avoid merge conflicts.
In JSON, an „object is an unordered set of name/value pairs“, but for the "tca"
fields the PHP ability of sorting hash tables is used. The "tca" fields may not
be sorted to keep the order of the column like the user selected it.