Fix TSConfig conditionals for hidden pages and content elements; #203
The default query restrictions in TYPO3 hide pages and content elements
that are hidden, or that have starttime and/or endtime set, and the
current time is outside of that time frame.

Mask uses a custom TSConfig conditional function that checks the
database for CType in case of tt_content, or backend_layout in case of
pages, to insert its own extended fields (pages) or set overridden
labels for inputs (tt_content).

This also adds a small static cache for the CType check to avoid
querying the database for each TSConfig conditional.