Remove setting the richtextConfiguration to default in tca
In TYPO3 v10 the order was changed in which richtext configuration is
overridden. Tca now has precedence over RTE.default.preset. Remove the
explicit setting of richtextConfiguration as default as this is the
default anyway.

Note: It is necessary to save the mask elements which have richtext
fields in order to apply this change.

Resolves: #284
Resolves: #281