[BUGFIX] Remove table definition of inline fields completely
When creating an inline (repeating) field in mask, a new table
definition is being generated in mask.json. If one later decides to
remove this inline field, an empty table definition is still there.
Although the database analyzer removes this table, exceptions are
raised when working with an existing content element, because there are
tca configurations for a non existing table.

Resolves: #265
Resolves: #255
Resolves: #272