[BUGFIX] Fix order of fields and tabs defined for backend layouts
The problem was that the order of showitem was taken from the mask tca
array instead of columns. This was necessary because backend layouts
can share fields and the order may vary. Now, instead of creating the
whole showitem string for all fields, we extend the showitem for the
specific page type and backend layout just before the tca is processed
with the help of a FormDataProvider. This way it is possible to use the
predefined order of fields and tabs.

Resolves: #214