[FEATURE] Add timestamp field
This feature adds the new field "Timestamp". Values in this field are
stored as integers (unix timestamps) as opposed to mysql date types.
Therefore the timezone is not converted back and forth and causes
wrong displaying of the date. This field should be used over the current
Date and Datetime fields.

Fixes: #234
Fixes: #260