[BUGFIX] Fix blindLinkOptions collision with same keys
If there is a key "tx_mask_link" for example and the same key is used
in a repeating element, then the blindLinkOptions were taken from the
first one.

To fix this we add the current parent as argument to the ViewHelper
instead of trying to resolve it from the element key.

Fixes: #397