[FEATURE] Fully migrate mask array structure to objects
For every array key elements, palettes, sql and tca there are now
appropriate models, which hold the configuration as public properties.
Any preprocessing of the configuration is done once, when
LoaderInterface->load() is called. The TableDefinition, which was
previously introduced, holds these different definitions as object
properties now, instead of arrays. The whole code was adapted to the
new object based API.