[BUGFIX] Use correct table and CType for adding relations
InlineHelper->getInlineElements used the field name as the table name
for the child relations. This does only work for inline fields, because
the field name and table are the same. Instead, the provided child
table name must be used.

The same goes for the cType. An inline field of type tt_content has
probably another cType than the parent. This can be fetched from the
data row.

Fixes: #439