[BUGFIX] Fix field key checking (for real now)
Field key checking has a tradition of not functioning
properly. It is more complicated than it seems: It is
for example possible to share keys between elements,
but some types must never be shared. These are inline
and content. On the other hand, fields in inline fields
can have the same key as elsewhere, because it is an
own table. Palettes, however, should not be treated
as parent fields. Except they are themselves in an
inline field. They also should not share same keys,
but their keys are now auto-generated, so no check
is needed.

You see, it tends to become more complicated, as
more you think about it. The Frontend and Backend
have also to harmonize.

These issues should now be addressed properly.

Related: #199