[BUGFIX] Fix breaking behaviour if ext-config is not defined.
Mask relied too heavy on existing extension
configuration. This patch does a couple of things:

- Loaders do not cause any errors, if extension
config is missing. Only if it is defined, but
wrong. The write method now also throws errors, if
persisting fails.

- Defining the template path is not strictly
necessary anymore. It is still possible to build
content elements, without having generated
templates.

- If errors with persisting happen while saving,
these are now caught and shown as Notifications.
Also the current configuration is not thrown away.

- Mask requires at least a configuration, where to
store its definitions. If this minimum requirement
is not met, a warning is shown in the module and
building new elements is not possible until valid
configuration is provided.

- Default values for extension configuration
removed. It does not make sense to provide a dummy
extension, which is not installed actively other
than for quick testing. Users should always
provide their own paths. This will also help us
later to remove the custom getAbsFileName method,
which allowed uninstalled extensions.

- Recommend LocalConfiguration.php over Additional
because the latter should not be versioned and
used for environment overrides.

Fixes: #429