[BUGFIX] Resolve old rte type format
Since the field type resolving refactoring, a
breaking change has been introduced for resolving
the rte type via element options.

This was good and needed, but the breaking change
should not be introduced in a minor version.

This patch adds a compatibility layer in the
JsonLoader, to be able to still resolve to rte.

(cherry picked from commit 7dcad6b4a13a7dd0122b5e762168ab6228d58779)