[BUGFIX] Always apply defaults when loading definition
Because TCA can change over time, the defaults
change as well. For this reason we need to always
apply the defaults from Defaults.php (tca_out).

tca_out means, these defaults can't be edited in
the backend and will always be used. Prior, these
would only be added once, when adding a new field.

To prevent breaking changes, they are now always
set.

A trait has been added, and can be used if it is
needed. Mask will use it for its Json loaders.
Custom loaders might not need it. If using it,
one must add an entry in the Services.yaml for
dependency injection to work with traits. See this
commit for reference how to do it.