[TASK] Clean up config when loading definition
Config cleanup is important, to get rid of old
configuration, which is not available anymore.
This happens of course after on-the-fly migration
is completed.

Prior to this, config cleanup was only done in the
FieldsController. The intent here was to only
present valid data to VueJS.

Impact: When saving or running the mask:convert
command, obsolete configuration is removed.