[BUGFIX] Preserve order of elements
In order to prepare a future sorting of elements,
a usort function was used together with a new
sorting key to order the elements.

However, this led to arbitrary order of elements
in PHP version less than 8. In php 8 a stable
sorting was introduced.

To prevent a changing order when updating, a
stable usort has been added, which takes effect
on lower php versions.