[BUGFIX] Fetch workspace ID from context API
In a project (11.5.x-dev, PHP 8.1) the
GLOBALS['BE_USER'] object is unset in the frontend
request resulting in a PHP error message when
calling the frontend with mask inline elements. In
newer TYPO3 versions, the corresponding properties
('workspace' in this case) should be fetched from
the Context API. That's what the change does.