[BUGFIX] Prevent empty TCA error for extension fields
In previous Mask versions, almost all fields were
allowed, which were found in tt_content TCA. This
was not a good idea, as a lot of fields have
special meanings and shouldn't be reused in Mask.

Of course, people used such fields, and now it
breaks, because Mask can't find any TCA for fields
loaded after Mask (if the cache is empty).

For the most part, the type is not that important
in the early stages of TCA generation. Thus, the
exception can be safely caught and a fall back to
type string can be done.

Also, all already added fields in the config are
now allowed in the field selector.

In the future, a configuration or better event can
be added, to dynamically add and remove allowed
fields.