[BUGFIX] Ignore disabled live records in workspace
In workspace mode, we need to first fetch the
original record uids. Because of this, we have to
disable the default restrictions in workspace mode
so the querybuilder won't filter them out.

Drive-by change:
Disabled items are not visible again in backend
preview, as it should be a preview for the
frontend. This was added as a drive-by change for
Mask v5.

In addition, the code has been cleaned up a bit.
- Added named parameters with query builder.
- Added variables for clearer understanding.

Fixes: #468
(cherry picked from commit 4117317bb61c5f178db4955addf30367ba74263e)