[BUGFIX] Resolve vendor package paths
Because of the overridden getAbsFileName method,
which will be finally removed in v8, the resolving
of vendor paths did not work.

A compatibility layer has been added to this
method, to still work in TYPO3 v10 and v11 in
addition to extensions, which are not loaded.

Fixes: #477

(cherry picked from commit 32e6821387afc5177dd88e28b3ecbb4cc0667520)