[FEATURE] Add module for handling item lists
A new module was added to conveniently manage
item lists for select, check and radio types.

In addition, the array sorting was moved into an
own class and is now always used, when calling the
toArray method of the TableDefinitionCollection.
This makes the StorageRepository slimmer and
ensures consistent sorting when e.g. using the
Loader in cli context.