[FEATURE] Implement #92942 - Allow icon overlay for elements
It is now possible to define an icon overlay for
the newContentElementWizard view. This can be
defined in the Mask settings beneath the standard
icon and color. It is optional and will only be
displayed, if an overlay icon is selected.

Note: It is possible to define a custom icon with
[element_key]_overlay.[png|svg]. But you have to
select an icon from the icon picker. It doesn't
matter which. This is a limitation of the current
implementation and will be hopefully resolved in
upcoming Mask versions.