[FEATURE] Allow to use core bodytext field as text
Since the beginning of this extension, the core
bodytext field was treated as richtext field. This
was hard coded and there was no way to use it as a
normal text field.

In order to make it possible to do so, the
TcaFieldDefinition is extended by the option
bodytextTypeByElement. This property holds the
type (text or richtext) for the bodytext field
depending on the element it is used in. Helper
methods were added, to get the appropriate type.

Now that the type can be distinguished, it has to
be selectable in the backend module. This is done
in the AjaxController in the method existingTca.
This method checks, if we are searching for a text
or richtext field and the core field bodytext is
in the list.

There is also a bugfix included:
The columnsOverrides for the bodytext field is now
only added, if the element uses this field. This
prevents unnecessary TCA generation.

Fixes: #408