[FEATURE] Introduce MigrationManager
The MigrationManager takes care of all necessary
migrations from older Mask versions. If a class
implements the MigrationInterface is will be
automatically tagged and passed to the
MigrationManager. A Migration defines, from what
version below it should migrate. The
RepeatableMigrationInterface can be used to tag
Migrations, that can be needed multiple times, if
something changes in the future.

The TableDefinitionCollection is extended with
a version property. It is used to indicate, in
which version the definition was stored last time.
This is used to decide, whether a Migration is
necessary or not.