[BUGFIX] Fix backend layout image path in TYPO3 v10
Since the swap to core FAL functionality, the
site prefix is gone in TYPO3 v10 for backend
layout preview icons.

In TYPO3 v11 all public URLs are changed to
absolute paths, so this doesn't affect newer
versions.

The change is simply what is done in the new
PublicUrlPrefixer since TYPO3 v11.

Fixes: #492