[BUGFIX] Fix persisting of shared array-like fields
With the Mask version 7.2 new modules were
introduced for items, groups and item sorting.

These modules also brought array configurations
into the game. This means that the old method of
plainly using mergeRecursiveWithOverrule doesn't
work anymore, as changes in arrays should be
completely replaced.

The most simple way is to first remove the old
configuration and override it with the new one.
This is now done, if the field is not removed. For
fields, which are not shared, this was done
already by completely removing the tca entry.

Fixes: #495