[BUGFIX] Prevent creation of fallback definition version
If an empty array is passed to
`TableDefinitionCollection::createFromArray`, then
it will fall back to a legacy version number. The
JsonSplitLoader now checks, if the definition is
not empty and only then passes the array to the
factory method.

This fixes an issue, where an empty definition
in JsonSplitLoader will always trigger a migration
notice.

Fixes: #502