[TASK] Migrate to use ES6 modules for TYPO3 v12
The source files are migrated to ES6,
and a ES6 bundle is created for TYPO3 v12,
and a AMD bundle for TYPO3 v11.

Indentation of javascript files is not adapted within
this patch to keep the differences small and reviewable.
That up for another patch.

npm install --save \
    vue@^2.6.0 \
    vuedraggable@^2.24.0 \
    @fonticonpicker/fonticonpicker@^3.1.0

npm install --save-dev \
    rollup@^3.2.0 \
    @rollup/plugin-node-resolve@^15.0.0 \
    @rollup/plugin-commonjs@^23.0.0 \
    @rollup/plugin-replace \
    rollup-plugin-minification@^0.2.0