[BUGIFX] Streamline InlineHelper->addIrreToData
This patch ensures, that the element definition is
preserved for the whole addIrreToData execution
stack. For this, the origin table needs to be
carried over all related methods as an additional
parameter.

Notable changes:
If no backend layout is found, the data array is
not filled anymore with all pages fields. This
makes absolutely no sense and should not be relied
on.

Related: #529