[BUGFIX] Prevent Mask file fields to be treated as core fields
TYPO3 v12 introduced the new TCA type="file". This is great and so Mask
removed the default TCA type="inline" and
foreign_table="sys_file_reference" for it.

The problem: Standard file fields with no option set have their config
array removed completely. This happens in TcaFieldDefinition in the
method removeBlankOptions. After the next save in the backend builder,
Mask assumes this field is a core field and sets coreField=1. This is
game over and needs to be fixed manually after.

To circumvent this problem, the fields "maxitems" and "minitems"
are kept in the config, eventhough they are empty and are added for
every file field, if not defined.

Fixes: #533