[BUGFIX] Prevent error when be:uri.editRecord is used in preview template (#549)
Without the request on the rendering context, the uri.editRecord and link.editRecord ViewHelpers of the backend module are unable to automatically generate a returnUrl.

In the case of uri.editRecord we get an exception in the page module, and with the link.editRecord we'll end up on a blank page, when trying to use the close button to return to the previous page.

Co-authored-by: Lars Peter Søndergaard <l.soendergaard@lichtflut-medien.de>
(cherry picked from commit 2ea97c75cdac3fe64ec9558890a4f24e1e74fab4)