[FEATURE] Implement columnsOverride (#559)
This feature adds a feature toggle to Mask to activate the usage of TCA columnsOverride.
With columnsOverride it is possible to have different TCA config for the same field per
element. This way it is even easier to reuse existing fields and customize the config
for a given element. This feature is disabled per default and needs to be enabled in the
extension settings. A migration command restructures your Mask config to the new format.
Core fields can be overriden even without the feature being active.

Co-authored-by: Sebastian Dernbach <sebastian.dernbach@ionos.com>
Co-authored-by: Nikita Hovratov <nikita.h@live.de>