[BUGFIX] Remove superfluous parent palette information in json-split
In json-split mode it is not necessary to keep all parent palette
information of every element in the installation. The current element
is sufficient and the array will be merged again when loaded.

This was not harmful before, but resulted in changes of every json file
which shares this field.

Fixes: #601
(cherry picked from commit 6db1849e643c6da025d34ac4978ee843daa7d076)